Remove old SelectCategoriesDialog
