+++ app/src/main/java/com/yurii/youtubemusic/ui/SelectCategoriesDialog.kt
@@ -5,57 +5,49 @@ import android.content.Context
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.yurii.youtubemusic.R
 import com.yurii.youtubemusic.models.Category
-import com.yurii.youtubemusic.utilities.ServiceLocator
 
-private typealias OnApplyCallBack = (categories: ArrayList<Category>) -> Unit
-
-class SelectCategoriesDialog private constructor(private val context: Context) {
-    private var callBack: OnApplyCallBack? = null
-    private lateinit var categories: List<Category>
-    private var alreadySelectedCategories: List<Category>? = null
+class SelectCategoriesDialog constructor(
+    private val context: Context,
+    private val allCategories: List<Category>,
+    private val alreadySelectedCategories: List<Category>,
+    private val onApplyCallBack: (List<Category>) -> Unit
+) {
     private val selectedCategories = ArrayList<Category>()
 
-    private fun create() {
-        val categoriesNames = categories.map { it.name }
-
+    fun show() {
         MaterialAlertDialogBuilder(context).apply {
             setTitle(R.string.label_playlists)
-            if (categories.isEmpty()) {
-                setView(R.layout.layout_no_categories)
-            } else {
-                setMultiChoiceItems(categoriesNames.toTypedArray(), getCheckedItems()) { _, which, isChecked ->
-                    categories.find { it.name == categoriesNames[which] }?.also { category ->
-                        if (isChecked)
-                            selectedCategories.add(category)
-                        else
-                            selectedCategories.remove(category)
-                    }
-                }
-                setPositiveButton(R.string.label_ok) { _, _ -> callBack?.invoke(selectedCategories) }
-                setNegativeButton(R.string.label_cancel, null)
-            }
+            setChoices(this)
+            setPositiveButton(R.string.label_ok) { _, _ -> onApplyCallBack.invoke(selectedCategories) }
+            setNegativeButton(R.string.label_cancel, null)
         }.show()
     }
 
-    private fun getCheckedItems(): BooleanArray? {
-        val checks = BooleanArray(categories.size)
-        categories.forEachIndexed { index, category ->
-            if (alreadySelectedCategories?.contains(category) == true) {
+    private fun setChoices(builder: MaterialAlertDialogBuilder) {
+        val categoriesNames = allCategories.map { it.name }
+        if (categoriesNames.isEmpty()) {
+            builder.setView(R.layout.layout_no_categories)
+        } else {
+            builder.setMultiChoiceItems(categoriesNames.toTypedArray(), getCheckedItems()) { _, which, isChecked ->
+                allCategories.find { it.name == categoriesNames[which] }?.also { category ->
+                    if (isChecked)
+                        selectedCategories.add(category)
+                    else
+                        selectedCategories.remove(category)
+                }
+            }
+        }
+    }
+
+    private fun getCheckedItems(): BooleanArray {
+        val checks = BooleanArray(allCategories.size)
+        allCategories.forEachIndexed { index, category ->
+            if (alreadySelectedCategories.contains(category)) {
                 selectedCategories.add(category)
                 checks[index] = true
             }
         }
 
         return checks
-    }
-
-    companion object {
-        fun selectCategories(context: Context, alreadySelectedCategories: List<Category>?, onApplyCallBack: OnApplyCallBack) {
-            SelectCategoriesDialog(context).apply {
-                callBack = onApplyCallBack
-                //this.categories =  ServiceLocator.providePreferences(context).getMusicCategories()
-                this.alreadySelectedCategories = alreadySelectedCategories
-            }.create()
-        }
     }
 }
